+++ packages/app-andrew/src/components/group-settings/Members/index.tsx
@@ -5,9 +5,12 @@ import {
   getGroupMembers,
 } from '@gtms/state-group'
 import { getDisplayName, getImage } from '@gtms/commons/helpers'
+import { Link } from '@gtms/commons/i18n'
 import { IGroup } from '@gtms/commons/models'
 // ui
+import { GoPlus } from 'react-icons/go'
 import { ErrorWrapper } from '@gtms/ui/ErrorWrapper'
+import { Modal } from '@gtms/ui/Modal'
 import { Spinner } from '@gtms/ui/Spinner'
 import { UserAvatar } from '@gtms/ui/UserAvatar'
 import styles from './styles.scss'
@@ -18,6 +21,7 @@ export const MembersSettings: FC<{ group: IGroup }> = ({ group }) => {
   const [state, setState] = useState<IGroupMembers>(
     groupMembersQuery.getGroupMembers()
   )
+  const [showInvite, setShowInvite] = useState(false)
 
   useEffect(() => {
     getGroupMembers(group.slug, 0, RECORDS_PER_PAGE)
@@ -46,17 +50,37 @@ export const MembersSettings: FC<{ group: IGroup }> = ({ group }) => {
           <p>No member, you should start a group promotion bitch</p>
         )}
 
+      {showInvite && (
+        <Modal
+          additionalStyles={styles.modalContent}
+          onClose={() => setShowInvite(false)}
+        >
+          Invite friends here
+        </Modal>
+      )}
+
       {!state.isLoading && !state.errorOccured && state.records.length > 0 && (
         <ul className={styles.membersList}>
           {state.records.map((user) => (
             <li className={styles.item} key={`group-admin-${user.id}`}>
-              <UserAvatar
-                image={getImage('200x200', user.avatar)}
-                size="100percent"
-              />
-              <span className={styles.name}>{getDisplayName(user)}</span>
+              <Link href={`/user/${user.id}`}>
+                <a>
+                  <UserAvatar
+                    image={getImage('200x200', user.avatar)}
+                    size="100percent"
+                  />
+                  <span className={styles.name}>{getDisplayName(user)}</span>
+                </a>
+              </Link>
             </li>
           ))}
+          <li className={styles.item} key="add-new-admin">
+            <button className={styles.btn} onClick={() => setShowInvite(true)}>
+              <i>
+                <GoPlus />
+              </i>
+            </button>
+          </li>
         </ul>
       )}
     </div>